put validator and validation_graph tests into their respective test modules
