+++ tests/validator/test_validation_graph.py
@@ -1,15 +1,26 @@
-from typing import Optional
+from typing import Any, Dict, Optional, Set, Tuple, Union
+from unittest import mock
 
+import pandas as pd
 import pytest
 
+import great_expectations.exceptions as ge_exceptions
+from great_expectations.core import IDDict
+from great_expectations.core.batch import Batch, RuntimeBatchRequest
 from great_expectations.core.expectation_configuration import ExpectationConfiguration
+from great_expectations.data_context.types.base import ProgressBarsConfig
+from great_expectations.execution_engine import PandasExecutionEngine
+from great_expectations.expectations.core import ExpectColumnValueZScoresToBeLessThan
+from great_expectations.expectations.registry import get_expectation_impl
 from great_expectations.validator.exception_info import ExceptionInfo
 from great_expectations.validator.metric_configuration import MetricConfiguration
 from great_expectations.validator.validation_graph import (
+    MAX_METRIC_COMPUTATION_RETRIES,
     ExpectationValidationGraph,
     MetricEdge,
     ValidationGraph,
 )
+from great_expectations.validator.validator import Validator
 
 
 @pytest.fixture
@@ -185,3 +196,297 @@ def test_ExpectationValidationGraph_get_exception_info(
 
     assert left_exception in exception_info
     assert right_exception in exception_info
+
+
+@pytest.mark.integration
+def test_parse_validation_graph():
+    df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, 6]})
+    expectation_configuration = ExpectationConfiguration(
+        expectation_type="expect_column_value_z_scores_to_be_less_than",
+        kwargs={
+            "column": "a",
+            "mostly": 0.9,
+            "threshold": 4,
+            "double_sided": True,
+        },
+    )
+    # noinspection PyUnusedLocal
+    expectation = ExpectColumnValueZScoresToBeLessThan(expectation_configuration)
+    # noinspection PyUnusedLocal
+    batch = Batch(data=df)
+    graph = ValidationGraph()
+    execution_engine = PandasExecutionEngine()
+    for configuration in [expectation_configuration]:
+        expectation_impl = get_expectation_impl(
+            "expect_column_value_z_scores_to_be_less_than"
+        )
+        validation_dependencies = expectation_impl(
+            configuration
+        ).get_validation_dependencies(configuration, execution_engine)
+
+        for metric_configuration in validation_dependencies["metrics"].values():
+            graph.build_metric_dependency_graph(
+                execution_engine=execution_engine,
+                metric_configuration=metric_configuration,
+                runtime_configuration=None,
+            )
+
+    ready_metrics, needed_metrics = graph._parse(metrics=dict())
+    assert len(ready_metrics) == 2 and len(needed_metrics) == 9
+
+
+# Should be passing tests even if given incorrect MetricProvider data
+@pytest.mark.integration
+def test_parse_validation_graph_with_bad_metrics_args():
+    expectation_configuration = ExpectationConfiguration(
+        expectation_type="expect_column_value_z_scores_to_be_less_than",
+        kwargs={
+            "column": "a",
+            "mostly": 0.9,
+            "threshold": 4,
+            "double_sided": True,
+        },
+    )
+    graph = ValidationGraph()
+    execution_engine = PandasExecutionEngine()
+    for configuration in [expectation_configuration]:
+        expectation_impl = get_expectation_impl(
+            "expect_column_value_z_scores_to_be_less_than"
+        )
+        validation_dependencies = expectation_impl(
+            configuration
+        ).get_validation_dependencies(
+            configuration,
+            execution_engine=execution_engine,
+        )
+
+        for metric_configuration in validation_dependencies["metrics"].values():
+            graph.build_metric_dependency_graph(
+                execution_engine=execution_engine,
+                metric_configuration=metric_configuration,
+                runtime_configuration=None,
+            )
+
+    # noinspection PyTypeChecker
+    ready_metrics, needed_metrics = graph._parse(metrics=("nonexistent", "NONE"))
+    assert len(ready_metrics) == 2 and len(needed_metrics) == 9
+
+
+@pytest.mark.integration
+def test_populate_dependencies():
+    df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, 6]})
+    expectation_configuration = ExpectationConfiguration(
+        expectation_type="expect_column_value_z_scores_to_be_less_than",
+        kwargs={
+            "column": "a",
+            "mostly": 0.9,
+            "threshold": 4,
+            "double_sided": True,
+        },
+    )
+    # noinspection PyUnusedLocal
+    expectation = ExpectColumnValueZScoresToBeLessThan(expectation_configuration)
+    # noinspection PyUnusedLocal
+    batch = Batch(data=df)
+    graph = ValidationGraph()
+    execution_engine = PandasExecutionEngine()
+    for configuration in [expectation_configuration]:
+        expectation_impl = get_expectation_impl(
+            "expect_column_value_z_scores_to_be_less_than"
+        )
+        validation_dependencies = expectation_impl(
+            configuration
+        ).get_validation_dependencies(
+            configuration,
+            execution_engine,
+        )
+
+        for metric_configuration in validation_dependencies["metrics"].values():
+            graph.build_metric_dependency_graph(
+                execution_engine=execution_engine,
+                metric_configuration=metric_configuration,
+            )
+
+    assert len(graph.edges) == 33
+
+
+@pytest.mark.integration
+def test_populate_dependencies_with_incorrect_metric_name():
+    df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, 6]})
+    expectation_configuration = ExpectationConfiguration(
+        expectation_type="expect_column_value_z_scores_to_be_less_than",
+        kwargs={
+            "column": "a",
+            "mostly": 0.9,
+            "threshold": 4,
+            "double_sided": True,
+        },
+    )
+    # noinspection PyUnusedLocal
+    expectation = ExpectColumnValueZScoresToBeLessThan(expectation_configuration)
+    # noinspection PyUnusedLocal
+    batch = Batch(data=df)
+    graph = ValidationGraph()
+    execution_engine = PandasExecutionEngine()
+    try:
+        graph.build_metric_dependency_graph(
+            execution_engine=execution_engine,
+            metric_configuration=MetricConfiguration(
+                "column_values.not_a_metric", IDDict()
+            ),
+        )
+    except ge_exceptions.MetricProviderError as e:
+        graph = e
+
+    assert isinstance(graph, ge_exceptions.MetricProviderError)
+
+
+@pytest.mark.integration
+def test_resolve_validation_graph_with_bad_config_catch_exceptions_true(
+    basic_datasource,
+):
+    df = pd.DataFrame({"a": [1, 5, 22, 3, 5, 10], "b": [1, 2, 3, 4, 5, None]})
+
+    batch = basic_datasource.get_single_batch_from_batch_request(
+        RuntimeBatchRequest(
+            **{
+                "datasource_name": "my_datasource",
+                "data_connector_name": "test_runtime_data_connector",
+                "data_asset_name": "IN_MEMORY_DATA_ASSET",
+                "runtime_parameters": {
+                    "batch_data": df,
+                },
+                "batch_identifiers": {
+                    "pipeline_stage_name": 0,
+                    "airflow_run_id": 0,
+                    "custom_key_0": 0,
+                },
+            }
+        )
+    )
+
+    expectation_configuration = ExpectationConfiguration(
+        expectation_type="expect_column_max_to_be_between",
+        kwargs={"column": "not_in_table", "min_value": 1, "max_value": 29},
+    )
+
+    runtime_configuration = {
+        "catch_exceptions": True,
+        "result_format": {"result_format": "BASIC"},
+    }
+
+    execution_engine = PandasExecutionEngine()
+
+    # noinspection PyUnusedLocal
+    validator = Validator(execution_engine=execution_engine, batches=[batch])
+
+    expectation_impl = get_expectation_impl(expectation_configuration.expectation_type)
+    validation_dependencies = expectation_impl().get_validation_dependencies(
+        expectation_configuration, execution_engine, runtime_configuration
+    )["metrics"]
+
+    graph = ValidationGraph()
+
+    for metric_configuration in validation_dependencies.values():
+        graph.build_metric_dependency_graph(
+            execution_engine=execution_engine,
+            metric_configuration=metric_configuration,
+            runtime_configuration=runtime_configuration,
+        )
+
+    metrics: Dict[Tuple[str, str, str], Any] = {}
+    aborted_metrics_info: Dict[
+        Tuple[str, str, str],
+        Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
+    ] = graph.resolve_validation_graph(
+        execution_engine=execution_engine,
+        metrics=metrics,
+        runtime_configuration=runtime_configuration,
+    )
+
+    assert len(aborted_metrics_info) == 1
+
+    aborted_metric_info_item = list(aborted_metrics_info.values())[0]
+    assert aborted_metric_info_item["num_failures"] == MAX_METRIC_COMPUTATION_RETRIES
+
+    assert len(aborted_metric_info_item["exception_info"]) == 1
+
+    exception_info = next(iter(aborted_metric_info_item["exception_info"]))
+    assert (
+        exception_info["exception_message"]
+        == 'Error: The column "not_in_table" in BatchData does not exist.'
+    )
+
+
+@mock.patch("great_expectations.validator.metric_configuration.MetricConfiguration")
+@mock.patch("great_expectations.validator.validation_graph.tqdm")
+@pytest.mark.unit
+def test_progress_bar_config_enabled(mock_tqdm, mock_metric_configuration):
+    execution_engine = PandasExecutionEngine()
+
+    # ValidationGraph is a complex object that requires len > 3 to not trigger tqdm
+    with mock.patch(
+        "great_expectations.validator.validation_graph.ValidationGraph._parse",
+        return_value=(
+            {},
+            {},
+        ),
+    ), mock.patch(
+        "great_expectations.validator.validation_graph.ValidationGraph.edges",
+        new_callable=mock.PropertyMock,
+        return_value=[
+            MetricEdge(left=mock_metric_configuration),
+            MetricEdge(left=mock_metric_configuration),
+            MetricEdge(left=mock_metric_configuration),
+        ],
+    ):
+        graph = ValidationGraph()
+        graph.resolve_validation_graph(
+            execution_engine=execution_engine,
+            metrics={},
+            runtime_configuration=None,
+        )
+
+    # Still invoked but doesn't actually do anything due to `disabled`
+    assert mock_tqdm.called is True
+    assert mock_tqdm.call_args[1]["disable"] is False
+
+
+@mock.patch("great_expectations.data_context.data_context.DataContext")
+@mock.patch("great_expectations.validator.metric_configuration.MetricConfiguration")
+@mock.patch("great_expectations.validator.validation_graph.tqdm")
+@pytest.mark.unit
+def test_progress_bar_config_disabled(
+    mock_tqdm, mock_metric_configuration, mock_data_context
+):
+    data_context = mock_data_context()
+    data_context.progress_bars = ProgressBarsConfig(metric_calculations=False)
+
+    execution_engine = PandasExecutionEngine()
+
+    # ValidationGraph is a complex object that requires len > 3 to not trigger tqdm
+    with mock.patch(
+        "great_expectations.validator.validation_graph.ValidationGraph._parse",
+        return_value=(
+            {},
+            {},
+        ),
+    ), mock.patch(
+        "great_expectations.validator.validation_graph.ValidationGraph.edges",
+        new_callable=mock.PropertyMock,
+        return_value=[
+            MetricEdge(left=mock_metric_configuration),
+            MetricEdge(left=mock_metric_configuration),
+            MetricEdge(left=mock_metric_configuration),
+        ],
+    ):
+        graph = ValidationGraph()
+        graph.resolve_validation_graph(
+            execution_engine=execution_engine,
+            metrics={},
+            runtime_configuration=None,
+            show_progress_bars=False,
+        )
+
+    assert mock_tqdm.called is True
+    assert mock_tqdm.call_args[1]["disable"] is True
